[ottlib, test_utils] Refactor the i2c bitbanging
Move the type Bit to the mod.rs to be used by other bitbangging libs.
