+++ sw/host/opentitanlib/src/test_utils/bitbanging/mod.rs
@@ -3,3 +3,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod i2c;
+
+#[repr(u8)]
+#[derive(Debug, PartialEq, Clone, Copy)]
+pub enum Bit {
+    Low = 0,
+    High = 1,
+}
+
+impl From<u8> for Bit {
+    fn from(val: u8) -> Self {
+        match val {
+            0x00 => Self::Low,
+            _ => Self::High,
+        }
+    }
+}
